Declarative pacman setup
This reads entries in .local/share/chezmoi/.chezmoidata/packages.yaml
and .config/chezmoi/chezmoi.yaml and compares them to the list of
currently *explicitly* installed packages. Missing packages are
installed as needed and marked as explicit. Extra pacakges are marked as
dependencies and then cleared up together in the final steps if no other
packages depend on them